backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
better-auth
pyjwt
passlib[bcrypt]
openai
mcp
dapr-ext-fastapi
